[refactor] 엔티티를 rdb 패키지로 이동 #4
