--- src/main/kotlin/dev/codeswamp/domain/article/entity/article/ArticleType.kt
@@ -1,6 +1,0 @@
-package dev.codeswamp.domain.article.entity.article
-
-enum class ArticleType {
-    DRAFT,
-    PUBLISHED,
-}
