+++ src/main/kotlin/dev/codeswamp/domain/article/entity/rdb/ArticleType.kt
@@ -1,0 +1,6 @@
+package dev.codeswamp.domain.article.entity.rdb.article
+
+enum class ArticleType {
+    DRAFT,
+    PUBLISHED,
+}
