+++ src/main/kotlin/dev/codeswamp/domain/article/entity/rdb/reaction/Comment.kt
@@ -1,6 +1,6 @@
-package dev.codeswamp.domain.article.entity.reaction
+package dev.codeswamp.domain.article.entity.rdb.reaction
 
-import dev.codeswamp.domain.article.entity.article.ArticleMetadata
+import dev.codeswamp.domain.article.entity.rdb.article.ArticleMetadata
 import dev.codeswamp.domain.user.entity.User
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
